// src/app/model/items.ts
export class Items {

    // Propiedades básicas

    id?: string;

    url: String;

    title: String;

    shortName?: String;

    ogTitle?: String;

    ogSaga?: String;

    description: string;

    genre1: string;
    genre2?: string;

    saga: string;
    sagaEntry: String;

    state: string;

    rating: number;
    fav?: boolean;

    // Propiedades videojuegos

    platform: String;

    company: String;

    duration: number;

    director?: String;




    
}